+++ 0.56.b/network/packet.cpp
@@ -375,7 +375,7 @@ void Packet::writeStringUNICODE(const char* value, bool terminate)
 	ASSERT(value != NULL);
 
 	WCHAR * buffer = reinterpret_cast<WCHAR *>(Str_GetTemp());
-	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, strlen(value));
+	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, static_cast<int>(strlen(value)));
 	
 	writeStringNUNICODE(buffer, terminate);
 #endif
@@ -407,7 +407,7 @@ void Packet::writeStringFixedUNICODE(const char* value, size_t size, bool termin
 	ASSERT(value != NULL);
 
 	WCHAR * buffer = reinterpret_cast<WCHAR *>(Str_GetTemp());
-	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, strlen(value));
+	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, static_cast<int>(strlen(value)));
 	
 	writeStringFixedNUNICODE(buffer, size, terminate);
 #endif
@@ -489,7 +489,7 @@ void Packet::writeStringNUNICODE(const char* value, bool terminate)
 	ASSERT(value != NULL);
 
 	WCHAR* buffer = reinterpret_cast<WCHAR *>(Str_GetTemp());
-	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, strlen(value));
+	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, static_cast<int>(strlen(value)));
 	
 	writeStringUNICODE(buffer, terminate);
 #endif
@@ -521,7 +521,7 @@ void Packet::writeStringFixedNUNICODE(const char* value, size_t size, bool termi
 	ASSERT(value != NULL);
 	
 	WCHAR* buffer = reinterpret_cast<WCHAR *>(Str_GetTemp());
-	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, strlen(value));
+	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), THREAD_STRING_LENGTH / sizeof(WCHAR), value, static_cast<int>(strlen(value)));
 	
 	writeStringFixedUNICODE(buffer, size, terminate);
 #endif
@@ -719,7 +719,7 @@ void Packet::readStringASCII(WCHAR* buffer, size_t length, bool includeNull)
 	
 	char* bufferReal = new char[length + 1];
 	readStringASCII(bufferReal, length, includeNull);
-	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), length, bufferReal, length + 1);
+	CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), static_cast<int>(length), bufferReal, static_cast<int>(length) + 1);
 	delete[] bufferReal;
 
 	// need to flip byte order to convert NUNICODE to UNICODE
@@ -773,7 +773,7 @@ void Packet::readStringUNICODE(char* buffer, size_t bufferSize, size_t length, b
 
 	WCHAR* bufferReal = new WCHAR[length + 1];
 	readStringNUNICODE(bufferReal, length, includeNull);
-	CvtNUNICODEToSystem(buffer, bufferSize, reinterpret_cast<NWORD *>(bufferReal), length + 1);
+	CvtNUNICODEToSystem(buffer, static_cast<int>(bufferSize), reinterpret_cast<NWORD *>(bufferReal), static_cast<int>(length) + 1);
 	delete[] bufferReal;
 #endif
 }
@@ -819,7 +819,7 @@ void Packet::readStringNUNICODE(char* buffer, size_t bufferSize, size_t length, 
 
 	WCHAR* bufferReal = new WCHAR[length + 1];
 	readStringUNICODE(bufferReal, length, includeNull);
-	CvtNUNICODEToSystem(buffer, bufferSize, reinterpret_cast<NWORD *>(bufferReal), length + 1);
+	CvtNUNICODEToSystem(buffer, static_cast<int>(bufferSize), reinterpret_cast<NWORD *>(bufferReal), static_cast<int>(length) + 1);
 	delete[] bufferReal;
 #endif
 }
@@ -855,7 +855,7 @@ size_t Packet::readStringNullASCII(WCHAR* buffer, size_t maxlength)
 
 	char* bufferReal = new char[maxlength + 1];
 	readStringNullASCII(bufferReal, maxlength);
-	long length = CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), maxlength, bufferReal, maxlength + 1);
+	long length = CvtSystemToNUNICODE(reinterpret_cast<NWORD *>(buffer), static_cast<int>(maxlength), bufferReal, static_cast<int>(maxlength) + 1);
 	delete[] bufferReal;
 
 	// need to flip byte order to convert NUNICODE to UNICODE
@@ -903,7 +903,7 @@ size_t Packet::readStringNullUNICODE(char* buffer, size_t bufferSize, size_t max
 
 	WCHAR* bufferReal = new WCHAR[maxlength + 1];
 	readStringNullNUNICODE(bufferReal, maxlength);
-	long length = CvtNUNICODEToSystem(buffer, bufferSize, reinterpret_cast<NWORD *>(bufferReal), maxlength + 1);
+	long length = CvtNUNICODEToSystem(buffer, static_cast<int>(bufferSize), reinterpret_cast<NWORD *>(bufferReal), static_cast<int>(maxlength) + 1);
 	delete[] bufferReal;
 #endif
 
@@ -943,7 +943,7 @@ size_t Packet::readStringNullNUNICODE(char* buffer, size_t bufferSize, size_t ma
 
 	WCHAR* bufferReal = new WCHAR[maxlength + 1];
 	readStringNullUNICODE(bufferReal, maxlength);
-	long length = CvtNUNICODEToSystem(buffer, bufferSize, reinterpret_cast<NWORD *>(bufferReal), maxlength + 1);
+	long length = CvtNUNICODEToSystem(buffer, static_cast<int>(bufferSize), reinterpret_cast<NWORD *>(bufferReal), static_cast<int>(maxlength) + 1);
 	delete[] bufferReal;
 #endif
 
@@ -1120,7 +1120,7 @@ void PacketSend::initLength(void)
 //	DEBUGNETWORK(("Packet %x starts dynamic with pos %ld.\n", m_buffer[0], m_position));
 
 	m_lengthPosition = m_position;
-	writeInt16(m_lengthPosition);
+	writeInt16(static_cast<WORD>(m_lengthPosition));
 }
 
 void PacketSend::fixLength()
@@ -1131,7 +1131,7 @@ void PacketSend::fixLength()
 
 		size_t oldPosition = m_position;
 		m_position = m_lengthPosition;
-		writeInt16(oldPosition);
+		writeInt16(static_cast<WORD>(oldPosition));
 		m_position = oldPosition;
 		m_lengthPosition = 0;
 		m_length = m_position;
